+++ apps/strapi-cms/src/api/web-page-summary/content-types/web-page-summary/lifecycles.ts
@@ -1,4 +1,4 @@
-import { upsertTypesenseCollection } from '@george-ai/typesense-client'
+import { ensureCollectionExists, upsertWebpageSummary } from '@george-ai/typesense-client'
 
 const transformAndUpsertSummary = async (id) => {
   const webPageSummaryResult = await strapi.entityService.findOne(
@@ -24,8 +24,8 @@ const transformAndUpsertSummary = async (id) => {
       ? 'published'
       : 'draft',
   }
-
-  upsertTypesenseCollection(webPageSummary)
+  await ensureCollectionExists()
+  await upsertWebpageSummary(webPageSummary)
 }
 
 export default {
