+++ apps/typesense-cli/src/drop.collection.ts
@@ -1,0 +1,7 @@
+import { deleteCollection } from '@george-ai/typesense-client'
+
+const dropCollection = async () => {
+  await deleteCollection('scraped_web_pages_summaries')
+}
+
+dropCollection()
